+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,73 @@
 # ai
 
+## 6.0.0-beta.108
+
+### Patch Changes
+
+- Updated dependencies [5dd4c6a]
+  - @ai-sdk/gateway@2.0.0-beta.59
+
+## 6.0.0-beta.107
+
+### Patch Changes
+
+- 8445d70: feat: export GatewayModelId and use to type LanguageModel
+
+## 6.0.0-beta.106
+
+### Patch Changes
+
+- Updated dependencies [1425df5]
+  - @ai-sdk/gateway@2.0.0-beta.58
+
+## 6.0.0-beta.105
+
+### Patch Changes
+
+- Updated dependencies [bca7e61]
+  - @ai-sdk/gateway@2.0.0-beta.57
+
+## 6.0.0-beta.104
+
+### Patch Changes
+
+- 2d166e4: feat(provider/gateway): add support for image models
+- Updated dependencies [2d166e4]
+  - @ai-sdk/gateway@2.0.0-beta.56
+
+## 6.0.0-beta.103
+
+### Patch Changes
+
+- Updated dependencies [cc5170d]
+  - @ai-sdk/gateway@2.0.0-beta.55
+
+## 6.0.0-beta.102
+
+### Patch Changes
+
+- Updated dependencies [5f66123]
+  - @ai-sdk/gateway@2.0.0-beta.54
+
+## 6.0.0-beta.101
+
+### Patch Changes
+
+- Updated dependencies [3782645]
+  - @ai-sdk/gateway@2.0.0-beta.53
+
+## 6.0.0-beta.100
+
+### Patch Changes
+
+- 8370068: fix(provider/google): preserve thoughtSignature through tool execution
+
+## 6.0.0-beta.99
+
+### Patch Changes
+
+- 384142c: feat(agent): add abortSignal parameter to generate and stream
+
 ## 6.0.0-beta.98
 
 ### Patch Changes