+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,70 @@
 
 ### Patch Changes
 
+- ai@6.0.0-beta.108
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [8445d70]
+  - ai@6.0.0-beta.107
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.106
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.105
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [2d166e4]
+  - ai@6.0.0-beta.104
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.103
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.102
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.101
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [8370068]
+  - ai@6.0.0-beta.100
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [384142c]
+  - ai@6.0.0-beta.99
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [b681d7d]
   - ai@6.0.0-beta.98